+++ tests/test-dirs/inlay-hint/samples.t
@@ -1,6 +1,6 @@
 Optional argument
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:0 \
   > -filename inlay.ml <<EOF
   > let f ?x () = x ()
   > EOF
@@ -20,7 +20,7 @@ Optional argument
 
 Optional argument with value
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:0 \
   > -filename inlay.ml <<EOF
   > let f ?(x = 1) () = x
   > EOF
@@ -40,13 +40,20 @@ Optional argument with value
 
 Labeled argument
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:0 \
   > -filename inlay.ml <<EOF
-  > let f ~x = x + 1
+  > let f ~x ~y:z = x + z
   > EOF
   {
     "class": "return",
     "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 13
+        },
+        "label": "int"
+      },
       {
         "pos": {
           "line": 1,
@@ -60,7 +67,7 @@ Labeled argument
 
 Case argument
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 2:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:0 \
   > -filename inlay.ml <<EOF
   > let f (Some x) = x + 1
   > EOF
@@ -78,9 +85,64 @@ Case argument
     "notifications": []
   }
 
+Pair arguments
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 2:0 \
+  > -filename inlay.ml <<EOF
+  > let f (a,b) = a+b
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 10
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 8
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+Record arguments
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 3:0 \
+  > -filename inlay.ml <<EOF
+  > type ('a, 'b) pair = {x:'a; y:'b}
+  > let f {x;y=b} = x+b
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 2,
+          "col": 12
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 2,
+          "col": 8
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
 Pattern variables without pattern-binding hint
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
   > -filename inlay.ml <<EOF
   > let f x =
   >   match x with
@@ -103,7 +165,7 @@ Pattern variables without pattern-binding hint
 
 Pattern variables with pattern-binding hint
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
   > -pattern-binding true \
   > -filename inlay.ml <<EOF
   > let f x =
@@ -135,7 +197,7 @@ Pattern variables with pattern-binding hint
 
 Let bindings without let hinting
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
   > -let-binding false \
   > -filename inlay.ml <<EOF
   > let f () = let y = 0 in y
@@ -146,10 +208,9 @@ Let bindings without let hinting
     "notifications": []
   }
 
-
 Let bindings with let hinting
 
-  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 -avoid-ghost-location false \
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
   > -let-binding true \
   > -filename inlay.ml <<EOF
   > let f () = let y = 0 in y
@@ -161,6 +222,276 @@ Let bindings with let hinting
         "pos": {
           "line": 1,
           "col": 16
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+Class-level let bindings without let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
+  > -let-binding false \
+  > -filename inlay.ml <<EOF
+  > class c x = let y = 1 in object method s = x + y end
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 9
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+Class-level let bindings with let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
+  > -let-binding true \
+  > -filename inlay.ml <<EOF
+  > class c x = let y = 1 in object method s = x + y end
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 1,
+          "col": 17
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 9
+        },
+        "label": "int"
+      }
+    ],
+    "notifications": []
+  }
+
+Top-level let bindings without let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
+  > -let-binding false \
+  > -filename inlay.ml <<EOF
+  > let y = 0
+  > EOF
+  {
+    "class": "return",
+    "value": [],
+    "notifications": []
+  }
+
+Top-level let bindings with let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 4:26 \
+  > -let-binding true \
+  > -filename inlay.ml <<EOF
+  > let y = 0
+  > EOF
+  {
+    "class": "return",
+    "value": [],
+    "notifications": []
+  }
+
+Binding operators without let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 5:0 \
+  > -let-binding false \
+  > -filename inlay.ml <<EOF
+  > let ( let* ) x f = match x with None -> None | Some x -> f x
+  > let ( let+ ) x f = match x with None -> None | Some x -> Some (f x)
+  > let ( and+ ) x y = let* x = x in let+ y = y in (x, y)
+  > let f x y = let+ x = x and+ y = y in x + y
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 4,
+          "col": 9
+        },
+        "label": "int option"
+      },
+      {
+        "pos": {
+          "line": 4,
+          "col": 7
+        },
+        "label": "int option"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 16
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 14
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 2,
+          "col": 16
+        },
+        "label": "'a -> 'b"
+      },
+      {
+        "pos": {
+          "line": 2,
+          "col": 14
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 16
+        },
+        "label": "'a -> 'b option"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 14
+        },
+        "label": "'a option"
+      }
+    ],
+    "notifications": []
+  }
+
+Binding operators with let hinting
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 5:0 \
+  > -let-binding true \
+  > -filename inlay.ml <<EOF
+  > let ( let* ) x f = match x with None -> None | Some x -> f x
+  > let ( let+ ) x f = match x with None -> None | Some x -> Some (f x)
+  > let ( and+ ) x y = let* x = x in let+ y = y in (x, y)
+  > let f x y = let+ x = x and+ y = y in x + y
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 4,
+          "col": 29
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 4,
+          "col": 18
+        },
+        "label": "int"
+      },
+      {
+        "pos": {
+          "line": 4,
+          "col": 9
+        },
+        "label": "int option"
+      },
+      {
+        "pos": {
+          "line": 4,
+          "col": 7
+        },
+        "label": "int option"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 39
+        },
+        "label": "'a"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 25
+        },
+        "label": "'a"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 16
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 3,
+          "col": 14
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 2,
+          "col": 16
+        },
+        "label": "'a -> 'b"
+      },
+      {
+        "pos": {
+          "line": 2,
+          "col": 14
+        },
+        "label": "'a option"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 16
+        },
+        "label": "'a -> 'b option"
+      },
+      {
+        "pos": {
+          "line": 1,
+          "col": 14
+        },
+        "label": "'a option"
+      }
+    ],
+    "notifications": []
+  }
+
+Support for @merlin.hide
+
+  $ $MERLIN single inlay-hints -start 1:0 -end 3:0 \
+  > -filename inlay.ml <<EOF
+  > let[@merlin.hide] f x = 2
+  > let f x = (fun y -> x+y+1) [@merlin.hide]
+  > EOF
+  {
+    "class": "return",
+    "value": [
+      {
+        "pos": {
+          "line": 2,
+          "col": 7
         },
         "label": "int"
       }