+++ CHANGES.md
@@ -6,6 +6,7 @@ unreleased
       (#1888)
     - `locate` can now disambiguate between files with identical names and contents
       (#1882)
+    - Improve Inlay Hints Handling with [@merlin.hide] and remove support for "avoid-ghost-location" (#1894)
     - `occurrences` now reports stale files (#1885)
   + ocaml-index
     - Improve the granularity of index reading by segmenting the marshalization